feat(#1): elasticsearch Entity 생성
