+++ src/main/java/com/neo/neomarket/entity/elasticsearch/AuctionLogEntity.java
@@ -1,0 +1,19 @@
+package com.neo.neomarket.entity.elasticsearch;
+
+import jakarta.persistence.Id;
+import lombok.Getter;
+import org.springframework.data.elasticsearch.annotations.Document;
+
+import java.time.LocalDateTime;
+
+
+@Getter
+@Document(indexName = "AuctionLog")
+public class AuctionLogEntity {
+    @Id
+    private Long id;
+
+    private Long bidPrice;
+
+    private LocalDateTime bidTime;
+}
