+++ src/main/java/com/neo/neomarket/entity/elasticsearch/UserLogEntity.java
@@ -1,0 +1,16 @@
+package com.neo.neomarket.entity.elasticsearch;
+
+import jakarta.persistence.Id;
+
+import java.time.LocalDateTime;
+
+public class UserLogEntity {
+    @Id
+    private Long id;
+
+    private String payType;
+
+    private String payStatus;
+
+    private LocalDateTime logTime;
+}
