+++ src/main/java/com/neo/neomarket/entity/mysql/AuctionPostEntity.java
@@ -1,4 +1,4 @@
-package com.neo.neomarket.entity;
+package com.neo.neomarket.entity.mysql;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
