+++ src/main/resources/application.properties
@@ -23,3 +23,8 @@ spring.jpa.properties.hibernate.format_sql=true
 
 
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+
+
+#elasticsearch ??
+spring.data.elas
+spring.data.elasticsearch.cluster-name=elasticsearch
